src/third_party/blink/web_tests/fast/forms/calendar-picker/month-picker-ax.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<script src="../../forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
<p id="description"></p>
<div id="console"></div>
<input type="month" id="month1" value="2000-01">

<script>
description('Tests if typing an arrow key dispatches |Focus| and |MarkDirty| a11y events.');

accessibilityController.setNotificationListener(function(axnode, type) {
    if (type == 'Focus') {
        debug('Focused');
    } else if (type == 'MarkDirty') {
        testPassed('Received MarkDirty');
        if (++markDirtyCounter == 2)
            finishJSTest();
    }
});
var markDirtyCounter = 0;
var month1 = document.getElementById('month1');
openPicker(month1, test1);

function test1() {
    eventSender.keyDown('ArrowRight');
}

</script>
</body>
</html>
