third_party/WebKit/LayoutTests/fast/forms/calendar-picker/month-picker-ax-expected.txt

Tests if typing an arrow key dispatches |Focus| and |ActiveDescendantChanged| a11y events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Focused: 
Focused: 
Focused: January 2000
PASS Received ActiveDescendantChanged
Focused: February 2000
PASS Received ActiveDescendantChanged
PASS successfullyParsed is true

TEST COMPLETE
